+++ bin/console
@@ -1,0 +1,8 @@
+#!/usr/bin/env ruby
+
+require "bundler/setup"
+require "markie"
+require "pry"
+
+Pry.start
+
