+++ lib/markie/version.rb
@@ -1,0 +1,3 @@
+module Markie
+  VERSION = "0.0.1"
+end
